# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "nadeemshaik1171998@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/addAddress": {
            "post": {
                "summary": "Add Address",
                "description": "User Can Add Address ",
                "operationId": "abd1c37c08074a7a38aead269f18a510",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "address",
                                    "city",
                                    "state",
                                    "landmark",
                                    "pincode",
                                    "address_type"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "landmark": {
                                        "type": "string"
                                    },
                                    "pincode": {
                                        "type": "integer"
                                    },
                                    "address_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address Added Successfully"
                    },
                    "401": {
                        "description": "Address alredy present for the user"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/updateAddress": {
            "post": {
                "summary": "Update Address",
                "description": "User Can Update Address ",
                "operationId": "70a93647591ea23a0125c8906cf88bc3",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id",
                                    "address",
                                    "city",
                                    "state",
                                    "landmark",
                                    "pincode",
                                    "address_type"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "state": {
                                        "type": "string"
                                    },
                                    "landmark": {
                                        "type": "string"
                                    },
                                    "pincode": {
                                        "type": "integer"
                                    },
                                    "address_type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Address Updated Successfully"
                    },
                    "401": {
                        "description": "Address not present add address first"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/deleteAddress": {
            "post": {
                "summary": "Delete Address",
                "description": " Delete Address ",
                "operationId": "7ee01710cba1151aafc2446826cede19",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book deleted Sucessfully from cart"
                    },
                    "404": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/getAddress": {
            "get": {
                "summary": "Get address ",
                "description": " Get Address ",
                "operationId": "957862060ee76d700293468a55f92cf5",
                "requestBody": {},
                "responses": {
                    "404": {
                        "description": "Address not found"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}